+++ core/core-common/src/context.rs
@@ -23,6 +23,7 @@ use crate::{KeyHandlingError, UserContext, UserDatabaseInitializer};
 use anyhow::{Error as AnyhowError, anyhow};
 use futures::TryFutureExt;
 use itertools::Itertools;
+use log_service::LogService;
 use proton_action_queue::action::{Action, WriterGuardError};
 use proton_action_queue::queue::{ActionError as QueueActionError, QueuedError};
 use proton_core_api::service::ApiServiceError;
@@ -49,7 +50,7 @@ use thiserror::Error;
 use tokio::sync::{Mutex, broadcast};
 use tokio::task::{JoinError, JoinHandle};
 use tokio_util::sync::CancellationToken;
-use tracing::{Level, debug, error, info, trace, warn};
+use tracing::{Level, error, info, warn};
 
 #[derive(Debug, Error)]
 pub enum CoreContextError {
@@ -246,7 +247,6 @@ pub struct Context {
     this: Weak<Self>,
     user_db_path: PathBuf,
     account_db_path: PathBuf,
-    log_path: Option<PathBuf>,
     account_stash: Stash,
     key_chain: Arc<dyn KeyChain>,
     user_db_initializers: Vec<Box<dyn UserDatabaseInitializer>>,
@@ -260,6 +260,7 @@ pub struct Context {
     on_session_deleted_broadcast: broadcast::Sender<(SessionId, UserId)>,
     pub event_poll_mode: EventPollMode,
     clock: CoreClock,
+    log_service: LogService,
 }
 
 const SESSION_OBSERVER_BROADCAST_CAPACITY: usize = 8;
@@ -294,7 +295,7 @@ impl Context {
         device_info_provider: Option<DynDeviceInfoProvider>,
         cache_path: impl Into<PathBuf>,
         connection_pool_size: Option<u32>,
-        log_path: Option<PathBuf>,
+        log_service: LogService,
         event_poll_mode: EventPollMode,
     ) -> CoreContextResult<Arc<Self>> {
         let initializers = initializers.into_iter().collect::<Vec<_>>();
@@ -329,7 +330,7 @@ impl Context {
             this: Weak::clone(this),
             user_db_path,
             account_db_path,
-            log_path,
+            log_service,
             key_chain,
             account_stash,
             user_db_initializers: initializers,
@@ -876,10 +877,10 @@ impl Context {
         let mut builder = ApiSession::builder().with_store(store);
 
         if app_settings.use_alternative_routing {
-            trace!("Using alternative routing");
+            info!("Using alternative routing");
             builder = builder.with_config(&self.api_config);
         } else {
-            debug!("Alternative routing setting is disabled");
+            info!("Alternative routing setting is disabled");
             builder = builder.with_config(self.api_config.clone().without_alternative_routing()?);
         }
 
@@ -971,8 +972,8 @@ impl Context {
         }
     }
 
-    pub fn get_log_path(&self) -> Option<&Path> {
-        self.log_path.as_deref()
+    pub fn log_service(&self) -> &LogService {
+        &self.log_service
     }
 
     /// Spawns a new task.
@@ -1153,7 +1154,7 @@ async fn on_session_deletion(
         tracing::debug!("Task received: {:?}", notifications);
         for notification in notifications {
             if let CoreSessionObserverNotification::Deleted(session_id, user_id) = notification {
-                tracing::debug!("User {user_id}'s session {session_id} has been deleted");
+                tracing::info!("User {user_id}'s session {session_id} has been deleted");
                 _ = hook_sender.send((session_id, user_id));
             }
         }