+++ mail/mail-common/examples/context_reuse.rs
@@ -1,6 +1,7 @@
 // cargo run --example context_reuse -- --username free --password free
 
 use clap::Parser;
+use log_service::LogService;
 use proton_core_api::session::Config;
 use proton_core_common::db::account::SessionEncryptionKey;
 use proton_core_common::event_loop::EventPollMode;
@@ -43,6 +44,10 @@ async fn main() {
         let keychain = InMemoryKeyChain::default();
         let key = SessionEncryptionKey::random();
         keychain.store(key).unwrap();
+        let config = log_service::Config::builder()
+            .name("log".into())
+            .directory(tmp_dir.path().into())
+            .build();
 
         MailContext::new(
             tmp_dir.path().join("session"),
@@ -55,7 +60,7 @@ async fn main() {
             Config::atlas(),
             None,
             None,
-            None,
+            LogService::new(config),
             EventPollMode::Manual,
         )
         .await