+++ mail/mail-common/examples/paginator.rs
@@ -1,4 +1,5 @@
 use clap::Parser;
+use log_service::LogService;
 use proton_core_api::services::proton::LabelId;
 use proton_core_api::session::Config;
 use proton_core_common::db::account::SessionEncryptionKey;
@@ -53,6 +54,10 @@ async fn main() {
     let keychain = InMemoryKeyChain::default();
     let key = SessionEncryptionKey::random();
     keychain.store(key).unwrap();
+    let config = log_service::Config::builder()
+        .name("log".into())
+        .directory(tmp_dir.path().into())
+        .build();
 
     let ctx = MailContext::new(
         tmp_dir.path().join("session"),
@@ -65,7 +70,7 @@ async fn main() {
         Config::default(),
         None,
         None,
-        None,
+        LogService::new(config),
         EventPollMode::Manual,
     )
     .await